fixes for non-numerical Index dtype
