+++ python/cudf/cudf/tests/test_series.py
@@ -762,8 +762,8 @@ def test_series_memory_usage():
     sliced_sr = sr[2:]
     assert sliced_sr.memory_usage() == 16
 
-    # sliced_sr[3] = None
-    # assert sliced_sr.memory_usage() == 80
+    sliced_sr[3] = None
+    assert sliced_sr.memory_usage() == 80
 
     sr = cudf.Series(["hello world", "rapids ai", "abc", "z"])
     assert sr.memory_usage() == 44
@@ -1196,47 +1196,3 @@ def test_explode(data, ignore_index, p_index):
             assert_eq(expect, got, check_dtype=False)
     else:
         assert_eq(expect, got, check_dtype=False)
-
-
-@pytest.mark.parametrize(
-    "arg",
-    [
-        1,
-        -1,
-        "b",
-        np.int32(1),
-        np.uint32(1),
-        np.int8(1),
-        np.uint8(1),
-        np.int16(1),
-        np.uint16(1),
-        np.int64(1),
-        np.uint64(1),
-    ],
-)
-def test_series_indexing(arg):
-    ps = pd.Series([1, 2, 3], index=pd.Index(["a", "b", "c"]))
-    gs = cudf.from_pandas(ps)
-
-    expect = ps[arg]
-    got = gs[arg]
-
-    assert_eq(expect, got)
-
-
-@pytest.mark.parametrize(
-    "dtype",
-    ["int32", "int16", "int8", "int64", "uint32", "uint16", "uint8", "uint64"],
-)
-def test_series_indexing_cudf_Scalar(dtype):
-    ps = pd.Series([1, 2, 3], index=pd.Index(["a", "b", "c"]))
-    gs = cudf.from_pandas(ps)
-
-    arg = cudf.Scalar(1, dtype=dtype)
-    got = gs[arg]
-
-    with pytest.raises(
-        TypeError, match=re.escape(f"'{arg}' is an invalid key"),
-    ):
-        expect = ps[arg]
-        assert_eq(expect, got)
